requirements.txt
flask
openai
anthropic
google-generativeai
requests
python-dotenv
gunicorn
flask_cors
